+++ examples/live/bitmex/bitmex_exec_tester.py
@@ -141,4 +141,5 @@ try:
     node.run()
 except KeyboardInterrupt:
     node.stop()
+finally:
     node.dispose()